requirements.txt
anthropic>=0.70.0
typer>=0.12.3
pydantic>=2.7.0
tqdm>=4.66.0
rich>=13.7.0
python-dotenv>=1.0.1
